+++ src/main/java/net/silentchaos512/gems/item/container/GemContainerScreen.java
@@ -1,0 +1,45 @@
+package net.silentchaos512.gems.item.container;
+
+import com.mojang.blaze3d.platform.GlStateManager;
+import net.minecraft.client.gui.screen.inventory.ContainerScreen;
+import net.minecraft.entity.player.PlayerInventory;
+import net.minecraft.util.ResourceLocation;
+import net.minecraft.util.text.ITextComponent;
+
+public class GemContainerScreen extends ContainerScreen<GemContainer> {
+    private static final ResourceLocation TEXTURE = new ResourceLocation("textures/gui/container/generic_54.png");
+
+    private final PlayerInventory playerInventory;
+    private final int inventoryRows;
+
+    public GemContainerScreen(GemContainer container, PlayerInventory playerInventory, ITextComponent title) {
+        super(container, playerInventory, title);
+        this.playerInventory = playerInventory;
+        this.inventoryRows = container.getInventoryRows();
+        this.ySize = 114 + this.inventoryRows * 18;
+    }
+
+    @Override
+    public void render(int mouseX, int mouseY, float partialTicks) {
+        this.renderBackground();
+        super.render(mouseX, mouseY, partialTicks);
+        this.renderHoveredToolTip(mouseX, mouseY);
+    }
+
+    @Override
+    protected void drawGuiContainerForegroundLayer(int mouseX, int mouseY) {
+        this.font.drawString(title.getFormattedText(), 8, 6, 4210752);
+        this.font.drawString(playerInventory.getDisplayName().getFormattedText(), 8, this.ySize - 96 + 2, 4210752);
+    }
+
+    @Override
+    protected void drawGuiContainerBackgroundLayer(float partialTicks, int mouseX, int mouseY) {
+        if (minecraft == null) return;
+        GlStateManager.color4f(1, 1, 1, 1);
+        minecraft.getTextureManager().bindTexture(TEXTURE);
+        int i = (this.width - this.xSize) / 2;
+        int j = (this.height - this.ySize) / 2;
+        blit(i, j, 0, 0, this.xSize, this.inventoryRows * 18 + 17);
+        blit(i, j + this.inventoryRows * 18 + 17, 0, 126, this.xSize, 96);
+    }
+}